Adopts new layout for Python node
