+++ nodes/script/script_python.py
@@ -1,0 +1,47 @@
+# -*- coding: utf-8 -*-
+###################################################################################
+#
+#  script_python.py
+#
+#  Copyright (c) 2022 Florian Foinant-Willig <ffw@2f2v.fr>
+#
+#  This program is free software; you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation; either version 2 of the License, or
+#  (at your option) any later version.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program; if not, write to the Free Software
+#  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
+#  MA 02110-1301, USA.
+#
+#
+###################################################################################
+from qtpy.QtWidgets import QSizePolicy, QTextEdit
+
+from core.nodes_conf import register_node
+from core.nodes_default_node import FCNNodeModel
+from nodes_locator import icon
+
+
+@register_node
+class Python(FCNNodeModel):
+
+    icon: str = icon("nodes_python_logo.png")
+    op_title: str = "Python"
+    op_category: str = "Script"
+    content_label_objname: str = "fcn_node_bg"
+
+    def __init__(self, scene):
+        super().__init__(scene=scene, inputs_init_list=[("In", True)], outputs_init_list=[("Out", True)])
+
+    def eval_operation(self, sockets_input_data: list) -> list:
+        code: str = "output_data = input_data\nprint(output_data)"
+        namespace = {'input_data': sockets_input_data[0], 'output_data': None}
+        exec(code, namespace)
+        return [namespace['output_data']]
